requirements.txt
tensorflow
pandas
numpy
pillow
pytesseract
scikit-learn
